requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
pytesseract>=0.3.10
paho-mqtt>=1.6.1
requests>=2.31.0
python-dotenv>=1.0.0
